requirements.txt
instagrapi>=2.0
python-dotenv>=1.0
pync>=2.0
requests>=2.0
python-dateutil>=2.0